+++ src/database/index.ts
@@ -1,7 +1,9 @@
+import { ObjectId } from 'mongodb'
 import * as Model from './Model'
-import ObjectId = require('./ObjectId')
 import * as connections from './connections'
 
-export { Model }
-export { ObjectId }
-export { connections }
+export {
+  ObjectId,
+  Model,
+  connections
+}