+++ src/exeptions/index.ts
@@ -1,5 +1,7 @@
 import * as HttpExeption from './HttpExeption'
 import * as Encapsulation from './Encapsulation'
 
-export { HttpExeption }
-export { Encapsulation }
+export {
+  HttpExeption,
+  Encapsulation
+}