+++ src/patches/index.ts
@@ -1,5 +1,7 @@
 import * as request from './request'
 import * as response from './response'
 
-export { request }
-export { response }
+export {
+  request,
+  response
+}