+++ src/routing/index.ts
@@ -2,6 +2,8 @@ import httpMethods from './httpMethods'
 import * as Router from './Router'
 import * as Route from './Route'
 
-export { httpMethods }
-export { Router }
-export { Route }
+export {
+  httpMethods,
+  Router,
+  Route
+}